pyproject.toml
[project]
name = "p8"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "asyncpg>=0.30",
    "cryptography>=44.0",
    "httpx>=0.28",
    "typer>=0.12",
    "pyyaml>=6.0",
    "fastembed>=0.4",
    "fastmcp>=2.0",
    "pydantic-ai[ag-ui]>=0.1",
    "sse-starlette>=2.0",
    "kreuzberg>=3.0",
    "pydub>=0.25",
    "authlib>=1.6",
    "pyjwt[crypto]>=2.8",
    "itsdangerous>=2.2.0",
    "hypercorn>=0.17",
]

[project.scripts]
p8 = "p8.api.cli:app"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["p8"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
